long_term/requirements.txt
pandas
pylint
pytest
python-dotenv
boto3
pymssql